+++ src/connection/mod.rs
@@ -1,108 +1,19 @@
 mod game_join;
+mod peer;
+pub mod shared;
 
-use crate::room;
+use crate::connection::{peer::Peer, shared::Shared};
 use futures::SinkExt;
-use std::collections::HashMap;
 use std::error::Error;
-use std::io;
 use std::net::SocketAddr;
 use std::sync::Arc;
 use tokio::net::TcpStream;
 use tokio_stream::StreamExt;
-use tokio::sync::{mpsc, Mutex};
+use tokio::sync::Mutex;
 use tokio_util::codec::{Framed, LinesCodec};
-
-/// Shorthand for the transmit half of the message channel.
-type Tx = mpsc::UnboundedSender<String>;
-
-/// Shorthand for the receive half of the message channel.
-type Rx = mpsc::UnboundedReceiver<String>;
-
-/// The state for each connected client.
-pub struct Peer<'a> {
-    /// The TCP socket wrapped with the `Lines` codec, defined below.
-    ///
-    /// This handles sending and receiving data on the socket. When using
-    /// `Lines`, we can work at the line level instead of having to manage the
-    /// raw byte operations.
-    pub lines: Framed<TcpStream, LinesCodec>,
-
-    /// Receive half of the message channel.
-    ///
-    /// This is used to receive messages from peers. When a message is received
-    /// off of this `Rx`, it will be written to the socket.
-    pub rx: Rx,
-
-    /// The room id that the peer belongs to.
-    pub room_id: u32,
-
-    /// The identifier uniquely identifying a user for their connection
-    /// TODO: make into a uuid
-    pub user_id: &'a str,
-}
-
-impl<'a> Peer<'a> {
-    /// Create a new instance of `Peer`.
-    pub async fn new(
-        state: Arc<Mutex<Shared>>,
-        lines: Framed<TcpStream, LinesCodec>,
-        room_id: u32,
-    ) -> io::Result<Peer<'a>> {
-        // Get the client socket address
-        let addr = lines.get_ref().peer_addr()?;
-
-        // Create a channel for this peer
-        let (tx, rx) = mpsc::unbounded_channel();
-
-        // Add an entry for this `Peer` in the shared state map.
-        let current_peer = state.lock().await.peer_count;
-        state.lock().await.peer_count += 1;
-        state.lock().await.peers.insert(addr, (tx, current_peer));
-
-        Ok(Peer {
-            lines,
-            rx,
-            room_id,
-            user_id: "billy",
-        })
-    }
-}
-
-/// Data that is shared between all peers in the chat server.
-///
-/// This is the set of `Tx` handles for all connected clients. Whenever a
-/// message is received from a client, it is broadcasted to all peers by
-/// iterating over the `peers` entries and sending a copy of the message on each
-/// `Tx`.
-pub struct Shared {
-    pub peers: HashMap<SocketAddr, (Tx, i32)>,
-    pub peer_count: i32,
-}
-
-impl Shared {
-    /// Create a new, empty, instance of `Shared`.
-    pub fn new() -> Self {
-        Self {
-            peers: HashMap::new(),
-            peer_count: 0,
-        }
-    }
-
-    /// Send a `LineCodec` encoded message to every peer, except
-    /// for the sender.
-    pub async fn broadcast(&mut self, sender: SocketAddr, message: &str) {
-        for peer in self.peers.iter_mut() {
-            if *peer.0 != sender {
-                //  && peer.1.1 == 1 {
-                let _ = peer.1 .0.send(message.into());
-            }
-        }
-    }
-}
 
 /// Process an individual chat client
 pub async fn process(
-    room_meta: Arc<Mutex<room::RoomMeta>>,
     state: Arc<Mutex<Shared>>,
     stream: TcpStream,
     addr: SocketAddr,
